+++ packages/abis/abis/IMapleLiquidityPositionParser.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Unlicense
-pragma solidity ^0.8.10;
+pragma solidity >=0.6.0 <0.9.0;
 
 interface IMapleLiquidityPositionParser {
     function parseAssetsForAction(address, uint256 _actionId, bytes memory _encodedActionArgs)