+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @enzymefinance/sdk
 
+## 0.0.24
+
+### Patch Changes
+
+- [#90](https://github.com/enzymefinance/sdk/pull/90) [`7e0c56c`](https://github.com/enzymefinance/sdk/commit/7e0c56c7f3c1840344e6b16739a5d361b8b6c56b) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add maple liquidity trade helpers
+
+- [#92](https://github.com/enzymefinance/sdk/pull/92) [`3a68713`](https://github.com/enzymefinance/sdk/commit/3a68713bd0c85a836a8e8e2f0385a5bc3ec57f70) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add the graph delegation trade helpers
+
+## 0.0.23
+
+### Patch Changes
+
+- [#86](https://github.com/enzymefinance/sdk/pull/86) [`8830633`](https://github.com/enzymefinance/sdk/commit/8830633b2ec5dff1e05b64bdd529f292ad659f75) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add compound v2 debt trade helpers
+
+- [#87](https://github.com/enzymefinance/sdk/pull/87) [`d5bce7a`](https://github.com/enzymefinance/sdk/commit/d5bce7a24ac56a373d02d632fafc2fa0ee011c66) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add convex voting trade helpers
+
 ## 0.0.22
 
 ### Patch Changes